+++ needle/modules/various/clean_storage.py
@@ -14,6 +14,12 @@ class Module(BaseModule):
     # ==================================================================================================================
     # UTILS
     # ==================================================================================================================
+    def __init__(self, params):
+        BaseModule.__init__(self, params)
+
+    def module_pre(self):
+        return BaseModule.module_pre(self, bypass_app=True)
+
     def clean_frida(self):
         self.printer.info("Cleaning Frida cache...")
         self.device.remote_op.dir_delete(Constants.DEVICE_PATH_FRIDA_CACHE, force=True)